Smooshish now has the full set of methods.
@bcclaywell: I have a situation where there are two classes that share
most of an interface: Smooshables implement a subset of the
functionality of SmooshableChains. In this commit I made their abstract
base class require all of the functionality of the SmooshableChain. We
want to be able to store either in SmooshableStacks, which seems to
imply that we would either need to do some dynamic casting, or something
like this.

Comments?